Do not hardcode "sid" as default for cowbuilder dist, support setting COWBUILDER_DIST [Closes #8]
On e.g. Ubuntu "sid" isn't supported, so we can't just hardcode it.
Instead support setting COWBUILDER_DIST via /etc/jenkins/debian_glue
and if it's not set then use "$distribution" if it's set, otherwise
default to the currently running distribution and just fallback to
"sid" if none of all of that worked.

Thanks: Daniel Thomas for the detailed bugreport