Drop the deprecated scripts generate-local-repository + include-in-repos
I do not want to support anything other than reprepro for
repository handling and if someone needs a simpler solution than
reprepro it should be a no-brainer anyways.

include-in-repos on the other side was the idea to split
build-and-provide-package into a build and repository only step
for scaling the build pipeline. Nowadays we also do support
BUILD_ONLY and PROVIDE_ONLY inside build-and-provide-package, so
there is no reason to duplicate that code any longer.

Stuff from include-in-repos which turned out to be useful and
if we should ever need it again it's available inside git log to
easier stumble upon it:
* reprepro includedsc
* iff section/priority is empty then reprepro will complain
  about "No section and no priority for" and error out in the
  cmdline above, therefore we retry with -S and -P being set
  ${SUDO_CMD:-} reprepro -v -S unstable -P extra -b "${REPOSITORY}" …