Always include $distribution in cow base name [closes #17]
This change might cause your system to create a new base.cow from
scratch. This doesn't cause any harm but increases disk space due
to the duplication. Though the change is useful because it makes
usage of $distribution much easier and also shows which
distribution you're actually building against.

To avoid the recreation of the base.cow run something like
(adjust "squeeze" for the distribution and "amd64" for the
architecture you're using):

|  cd /var/cache/pbuilder
|  mv base-amd64.cow base-squeeze-amd64.cow

Thanks: Christoph Berg for the suggestion