incoming repos: allow all kinds of distributions listed in packages enter the repos
If we're building e.g. a package which lists squeeze-backports as
its distribution then we still might want this package to have in
a distribution which usually provides packages built using
unstable. So replace the "Allow: unstable>$release" mapping with
"Default: unstable" to get that.

Thanks: Bernhard R. Link <brlink@debian.org> for the hint regarding "Default:"