Allow removal of packages from releases
By default, if building a package for a release, it is assumed, that
this version has never been part of the release. Thus building the same
version again will fail, since reprepro will not allow adding this.
This patch adds a new option REMOVE_FROM_RELEASE which will force a
removal of the package from a release prior to incoming processing,
which may be useful in some cases (e.g. wanted rebuilds of releases)