Support providing a custom pbuilder configuration file via PBUILDER_CONFIG
If the PBUILDER_CONFIG configuration is pointing to a file then
its argument is used for pbuilder's --configfile option.

For example when using the following option in /etc/jenkins/debian_glue:

  PBUILDER_CONFIG=/etc/jenkins/pbuilderrc

and the following configuration is present in /etc/jenkins/pbuilderrc:

  CCACHEDIR=/var/cache/pbuilder/ccache

then jenkins-debian-glue and its cowdancer system will use the
provided directory as ccache to speed up builds.

Development sponsored by Sipwise GmbH.