piuparts_wrapper: only expand scriptsdir when it is set
We started quoting $scriptsdir in 841b4b2 in case it is being passed a
string containing spaces.  That caused an issue whenever the scriptsdir
is empty: piuparts would be given an empty string as an argument which
does not play well.

This script makes $scriptdir to only contains the path, and expand it to
"--scriptsdir $scriptsdir" only if $scriptsdir exist.  That would let
the wrapper works properly whenever the scriptdir is empty by simply
discarding the option, but still quoting it properly when it is set.

Change-Id: I48c8e86ac007108ca6b650673087160a99d18a21