Make it possible to verify our own reprepro repositories by providing a pubkey.
Verifying the integrity of our repository is basically mandatory when the
build server is separated from the repository server. Additionally, if you
want to build a package for Squeeze that has a dependency on a package from
our repository the package list *MUST* be verified because "[trusted=yes]" in
the sources list is not yet supported by that apt version.