Allow overriding the source package dist
When DIST is set in a source job, generate-git-snapshot now uses that
dist in the debian/changelog. As this is only useful if one keeps the
built .changes file, this introduces a new setting KEEP_SOURCE_CHANGES
which must be set as well.

In the binary job configuration, one can get to the dist like this:

  export REPOS=$(awk '/^Distribution:/ { gsub("Distribution: ", ""); print $0 }' tmpdir/*_source.changes)

Also, in the binary job one should remove the source changes file, ex:

  rm -f tmpdir/*_source.changes

I'm using this successfully to build per-branch dists, in a development
environment where work happens on multiple repositories with shared
branch names.

With changes from @mika to turn the fatal error into a warning.