Support Gerrit integration without direct connection to Gerrit events
If the *-source job gets triggered directly via Gerrit then we
already have the according commit in $GIT_COMMIT and there's
nothing further special to do (we already supported this use case
since commit 017756fb40).

But if Gerrit does NOT directly trigger the *-source job then
we're usually not standing on the according $GIT_COMMIT. Also the
Gerrit patchset/review change possibly hasn't been even fetched
yet into the Git checkout/workspace. In $GIT_URL we have the
repository URL and rely on being able to fetch from there, using
the "changes" refs. (If someone has the need to make $GIT_URL
even more flexible this shouldn't be a problem, but until someone
actually has the needs I'm following KISS.)

Then we assume that $GERRIT_EVENT_TYPE is unset (which indicates
that it's not directly triggered via Gerrit) and
$GERRIT_PATCHSET_REVISION is set (and points to the according Git
commit that gets fetched).

To set this up all you need to do is triggering the *-source from
your *-gerrit job (or whatever the name of the Jenkins job that
receives the Gerrit events is) with parameter
GERRIT_PATCHSET_REVISION=$GERRIT_PATCHSET_REVISION and then the
*-source job does the rest.

Development sponsored by Sipwise GmbH