generate-git-snapshot: add DCH_LOG_RESTRICTION
DCH_LOG_RESTRICTION can be used to pass additional (path restriction)
parameters to git-dch. For example to restrict the repository path that
is looked at.

A path is for example required if debian is not a directory but a link to a
sub directory created during source package build (e.g. with
PRE_DCH_HOOK). In such a case the changelog would always
contain all commits of the repository if the path wasn't restricted.

Note: An extra parameter was chosen because DCH_OPT couldn't be used
since it isn't the last argument and overriding DCH_EXTRA_OPTS is also
not easily possible if the adjusted VERSION_STRING should be used.