generate-svn-snapshot: always generate changes file
If the debian_only variable is unset (because the svn property
'mergeWithUpstream' is unset) we use dpkg-source instead of
svn-buildpackage and then we don't have an according changes
file. If KEY_ID is set though we sign the changes file and expect
to have such a file present. Generate it through an according
dpkg-genchanges call.

Thanks: Christian Schneider <christian@chrisn.dtdns.net> for the bug report and patch