generate-svn-snapshot: support $SBP_OPTS to control svn-buildpackage behaviour + add -nc to its default options
Instead of hardcoding the options to be used with svn-buildpackage
let's support the $SBP_OPTS environment variable, similar to what
we do with $GBP_OPTS in generate-git-snapshot.

We also use `-nc` ("Do not clean the source tree") in
generate-git-snapshot's default $GBP_OPTS, so it makes sense to
use that also in the defaults of generate-svn-snapshot.

Thanks: Christian Schneider <christian@chrisn.dtdns.net> for the patch