Use adt-run's --output-dir option if available
autopkgtest >=3.16 no longer provides the --tmp-dir option, therefore
builds are failing with:

| + adt-run --tmp-dir /tmp/adt-15150//out --summary /tmp/adt-15150//summary [...]
| adt-run: error: unrecognized arguments: --tmp-dir /tmp/adt-15150//out

The --tmp-dir option was provided for backwards compatibility
reasons in adt-run.  Now that this option is gone in recent
versions of autopkgtest let's use the --output-dir option
instead.  To avoid breaking any setups with very old autopkgtest
where this option *might* not be present yet let's check for its
presence and fall back to --tmp-dir otherwise.

Thanks: Axel Beckert for the hint