Remove ARCH variable from cowbuilder command
Build of Redis package fails because ARCH is used in Makefile, with a
different meaning.
Removing it from cowbuilder allow Redis to be build and, as far as I saw,
does not break any other package.

Since use of ARCH with cowbuilder is optional, I found it safer just to remove
it from command line instead of changing more logic in build-and-provide script.