Add an option for enforcing Git commit in package version
This applies only in cases where the debian/changelog is explicitly
set to "UNRELEASED" which makes the generate-git-snapshot use plain
'dch' instead of 'gbp dch' or 'git-dch' which in turn passes the
generated version string directly as the new package version.

This is implemented as an additional option to avoid breaking
existing users which are just happy with the current behaviour.