Add variable to change changelog path for git-dch
In cases where the source package comes already with a debian folder which is located in some subfolder, it can be needed to specify the path to that file explicitly for git-dch to work correctly. That's why I introduced the variable "$DCH_CHANGELOG_FILE", so I could specify this file. It's empty and uninitialized if not specified, in which case the changelog is read from ./debian/changelog.