puppet: further changes to work with recent puppet versions
If ec2_public_ipv4 isn't set it warns with:

| Warning: Unknown variable: 'ec2_public_ipv4'. at

The file definition for the /var/lib/jenkins/jobs/jenkins-debian-glue-piuparts/
directory needs to include the trailing slash, otherwise the ressource isn't
properly detected in recent Puppet versions.

The "$PropertyList" inside the XML configuration needs to be escaped,
otherwise it's failing hard in recent Puppet versions.

Verified on:
* Ubuntu/xenial with puppet v3.8.5-2ubuntu0.1,
* Debian/stretch with puppet v4.8.2-5,
* Debian/jessie with puppet v3.7.2-4+deb8u1 (inside Virtualbox VM and on EC2)