Extract lintian to junit conversion to a new script
scripts/lintian-junit-report first runs lintian on the build host and
then post process the output to generate a JUnit file.

I have the use case to run lintian via a pbuilder hook to use the
lintian version that comes with the targetted distribution. The result
is a lintian.txt which I need to convert to JUnit.

Factor out the code that convert to JUnit into a new standalone script:
lintian2junit

Keep all options alike and pass them to the new script. So this should
be 100% back compatible.

Change-Id: I50ec91dae7088c284999d27cf9a234687332b83f