autopkg_binary variable needs to always be set
If you used ADT_OPTIONS, this would not allow
autopkg_binary variable to be set thus running
into an unbound variable error on line 591.
This separates the concerns of the options being
used and the actual binary being executed.

Also, adt-run and autopkgtest are not API compatible.
For instance, adt-run expects --changes to point
to the changes file whereas autopkgtest expects
the changes file pass as a positional arg. This
change takes that into consideration as well.