lintian-junit-report: fix for xunit 2.x Jenkins plugin
When using the Jenkins Xunit plugin, the generated XML lintian report
does not validate the plugin embeded XSD:

  Attribute 'name' must appear on element 'testsuite'.

The reporter has been made to mostly mimic the phpunit output but missed
the namel attribute. Add a dummy 'lintian' name to the testsuite
element.

The Jenkins plugin has to be configured to use phpunit-4.0.

One can then validate it using the embedded XSD:

 scripts/lintian-junit-report foo.dsc > lintian.xml
 git clone https://github.com/jenkinsci/xunit-plugin
 xmlllint --noout \
  --schema ./xunit-plugin/src/main/resources/org/jenkinsci/plugins/xunit/types/phpunit-4.0.xsd \
  lintian.xml

Add a note refering to the XSD provided by Jenkins.

Reference:
https://phabricator.wikimedia.org/T194096#6172999